# server/src/main/resources/openapi/family_accounting_openapi.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  description: |
    家庭记账系统的 RESTful API，支持用户认证、人员管理、账户管理、交易记录和统计。
    支持银行、微信、支付宝等多种账号，每种账号下可有多张卡或子账户。
    角色权限：
    - 管理员：访问所有数据，管理用户、账号、子账户。
    - 普通用户：仅访问通过 user_sub_accounts 关联的子账户和自己的交易/统计。
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
...
paths:
...
  /login:
    post:
      tags:
        - auth
        - Auth
      summary: 用户登录
      operationId: authLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 无效的用户名或密码
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /statistics:
    get:
      tags:
        - Statistics
      operationId: statisticsGet
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: accountId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: subAccountId
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: period
          in: query
          required: true
          schema:
            type: string
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: month
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: week
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: day
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
  /statistics/statistics:
    get:
      tags:
        - Statistics
        - statistics
      summary: 获取收支统计
      operationId: statisticsGet_1
      parameters:
        - name: period
          in: query
          description: 统计周期（月度或年度）
          required: true
          schema:
            type: string
            enum:
              - monthly
              - annual
        - name: year
          in: query
          description: 年份
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: query
          description: 按用户过滤（管理员专用，普通用户自动使用 JWT userId）
          required: false
          schema:
            type: integer
            format: int32
        - name: accountId
          in: query
          description: 按顶级账号过滤
          required: false
          schema:
            type: integer
            format: int32
        - name: subAccountId
          in: query
          description: 按子账户过滤
          required: false
          schema:
            type: integer
            format: int32
        - name: month
          in: query
          description: 月份（月度统计时需要）
          required: false
          schema:
            maximum: 12
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
        '400':
          description: 无效参数（参数格式错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
components:
  schemas:
...
    ApiError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          enum:
            - PERMISSION_DENIED
            - DUPLICATE_ACCOUNT
            - INVALID_TRANSACTION
            - DATA_NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - INVALID_REQUEST
            - VALIDATION_FAILED
            - INVALID_CREDENTIALS
        message:
          type: string
...
    LoginResponse:
      required:
        - role
        - token
      type: object
      properties:
        token:
          type: string
          description: JWT 令牌
        role:
          type: string
          description: 用户角色
          enum:
            - ADMIN
            - USER
    LoginRequest:
      required:
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 100
          minLength: 0
          type: string
...
    StatisticsResponse:
      required:
        - period
        - totalExpense
        - totalIncome
        - year
      type: object
      properties:
        period:
          type: string
          description: '统计周期：daily, weekly, monthly, yearly'
          enum:
            - daily
            - weekly
            - monthly
            - yearly
        year:
          maximum: 2100
          minimum: 2000
          type: integer
          description: 年份
          format: int32
        month:
          maximum: 12
          minimum: 1
          type: integer
          description: 月份（月度统计时需要）
          format: int32
        week:
          maximum: 53
          minimum: 1
          type: integer
          description: 周次（周度统计时需要）
          format: int32
        day:
          maximum: 31
          minimum: 1
          type: integer
          description: 日期（日度统计时需要）
          format: int32
        incomeByType:
          type: array
          description: 按类型统计的收入
          items:
            $ref: '#/components/schemas/StatisticsResponseIncomeByType'
        expenseByType:
          type: array
          description: 按类型统计的支出
          items:
            $ref: '#/components/schemas/StatisticsResponseExpenseByType'
        totalIncome:
          pattern: '^\d+(\.\d{2})?$'
          type: string
          description: 总收入（固定 2 位小数）
        totalExpense:
          pattern: '^\d+(\.\d{2})?$'
          type: string
          description: 总支出（固定 2 位小数）
    StatisticsResponseExpenseByType:
      required:
        - amount
        - typeId
        - typeName
      type: object
      properties:
        typeId:
          type: integer
          description: 交易类型 ID
          format: int32
        typeName:
          maxLength: 50
          minLength: 0
          type: string
          description: 交易类型名称（如“日常消费”）
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
          description: 支出金额（固定 2 位小数）
      description: 按类型统计的支出
    StatisticsResponseIncomeByType:
      required:
        - amount
        - typeId
        - typeName
      type: object
      properties:
        typeId:
          type: integer
          description: 交易类型 ID
          format: int32
        typeName:
          maxLength: 50
          minLength: 0
          type: string
          description: 交易类型名称（如“工资”）
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
          description: 收入金额（固定 2 位小数）
      description: 按类型统计的收入